refactor: reorganize the ananlysis module
